requirements.txt
python-pptx
requests
pyyaml
Pillow
pandas